feature(BreakableObject) : Add breakable object

+++ Assets/Script/IInteractable.cs
@@ -4,5 +4,5 @@ public interface IInteractable
     /// Interact with object
     /// </summary>
     /// <returns>Copper to return after interaction</returns>
-    public int Interact();
+    public float Interact();
 }
